requirements.txt
tensorflow
pillow
numpy
requests
gradio
flask
flask-cors
gunicorn